refactor: update component to use types
