+++ webpack.config.js
@@ -18,6 +18,7 @@ module.exports = {
         alias: {
             ...defaultConfig.resolve.alias,
             '@givewp/forms/types': path.resolve(__dirname, 'src/NextGen/DonationForm/resources/types.ts'),
+            '@givewp/forms/propTypes': path.resolve(__dirname, 'src/NextGen/DonationForm/resources/propTypes.ts'),
         },
     },
     entry: {
@@ -28,7 +29,7 @@ module.exports = {
             'NextGen/Gateways/Stripe/NextGenStripeGateway/nextGenStripeGateway.tsx'
         ),
         nextGenTestGateway: srcPath('NextGen/Gateways/NextGenTestGateway/nextGenTestGateway.tsx'),
-        classicTemplateJs: srcPath('NextGen/DonationForm/FormTemplates/ClassicFormTemplate/js/template.js'),
+        classicTemplateJs: srcPath('NextGen/DonationForm/FormTemplates/ClassicFormTemplate/js/template.ts'),
         classicTemplateCss: srcPath('NextGen/DonationForm/FormTemplates/ClassicFormTemplate/css/template.scss'),
     },
 };
